Restructing files to move help structs into parent
